升级引用库版本

+++ extend/pull-kline-mysql.go
@@ -43,7 +43,7 @@ func (this *PullKlineMysql) Name() string {
 }
 
 func (this *PullKlineMysql) Run(ctx context.Context, m *tdx.Manage) error {
-	limit := chans.NewWaitLimit(uint(this.Config.Limit))
+	limit := chans.NewWaitLimit(this.Config.Limit)
 
 	//1. 获取所有股票代码
 	codes := this.Config.Codes
